Handle BAG housenumber notational shortcomings
The BAG was designed on the assumption that housenumbers (ignoring
extensions) are always numeric. This is not always true in reality, so
in order to store housenumbers starting with a letter, a workaround was
devised where the letter prefix was appended to the street name instead.

In OSM this is not necessary. This commit adds a way to undo this
BAG-hack, and get the actual addresses for selected postcodes.