Replace term Parcel with Plot
We used to refer to BAG Ligplaats and Standplaats entities with the term
'Parcel'. However, the term 'Plot' is more appropriate and in line with
the use of place=plot and plot=static_caravan / plot=house_boat in OSM.
therefore, we refactored class and property names from Parcel to Plot.